requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
python-dotenv==1.0.1
msal==1.31.0
requests==2.32.3
openai==1.57.0
